feat: add single part time hook
