+++ src/cow-react/modules/advancedOrders/containers/DeadlineSelector/index.tsx
@@ -1,15 +1,13 @@
-import { Menu } from '@reach/menu-button'
-import { ordersDeadlines } from './deadlines'
 import { useCallback, useRef, useState } from 'react'
-import { ChevronDown } from 'react-feather'
-import { Trans } from '@lingui/macro'
+
 import { useDisplayDeadline } from './hooks/useDisplayDeadline'
 import { CustomDeadline } from './types'
-import QuestionHelper from 'components/QuestionHelper'
-import * as styledEl from './styled'
 import { CustomDeadlineSelector } from '../CustomDeadlineSelector'
 import { useDeadline } from '../DeadlineSelector/hooks/useDeadline'
 import { useUpdateDeadline } from './hooks/useUpdateDeadline'
+import { DeadlineDisplayPart } from '../DeadlineDisplayPart'
+import { Selector } from './pure/Selector'
+import * as styledEl from './styled'
 
 export function DeadlineSelector() {
   const { customDeadline } = useDeadline()
@@ -41,32 +39,15 @@ export function DeadlineSelector() {
 
   return (
     <styledEl.Wrapper>
-      <styledEl.Label>
-        <styledEl.LabelWrapper>
-          <span>Total time</span>
-          <QuestionHelper text={<Trans>This is some text here</Trans>} />
-        </styledEl.LabelWrapper>
-      </styledEl.Label>
+      <Selector
+        currentDeadlineNode={currentDeadlineNode}
+        selectDeadline={selectDeadline}
+        openModal={openModal}
+        customDeadline={customDeadline}
+        displayDeadline={displayDeadline}
+      />
 
-      <Menu>
-        <styledEl.Current ref={currentDeadlineNode as any} $custom={!!customDeadline}>
-          <span>{displayDeadline}</span>
-          <ChevronDown size="18" />
-        </styledEl.Current>
-
-        <styledEl.ListWrapper>
-          {ordersDeadlines.map((item) => (
-            <li key={item.value}>
-              <styledEl.ListItem onSelect={() => selectDeadline(item.value)}>
-                <Trans>{item.title}</Trans>
-              </styledEl.ListItem>
-            </li>
-          ))}
-          <styledEl.ListItem onSelect={openModal}>
-            <Trans>Custom</Trans>
-          </styledEl.ListItem>
-        </styledEl.ListWrapper>
-      </Menu>
+      <DeadlineDisplayPart />
 
       <CustomDeadlineSelector
         selectCustomDeadline={selectCustomDeadline}
